Init : config and jwt middleware added.

+++ model/basemodel.go
@@ -6,7 +6,7 @@ import (
 )
 
 type BaseModel struct {
-	ID        uint `gorm:"primaryKey"`
+	ID        int64 `gorm:"primaryKey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt gorm.DeletedAt `gorm:"index"`
